+++ src/stores/playingList.ts
